+++ api/src/main/java/mb/resource/classloader/ClassLoaderResourceLocations.java
@@ -1,20 +1,22 @@
 package mb.resource.classloader;
 
+import mb.resource.ReadableResource;
 import mb.resource.fs.FSResource;
+import mb.resource.hierarchical.HierarchicalResource;
 import org.checkerframework.checker.nullness.qual.Nullable;
 
 import java.io.Serializable;
 import java.net.URL;
 import java.util.ArrayList;
 
-public class ClassLoaderResourceLocations implements Serializable {
-    public final ArrayList<FSResource> directories;
-    public final ArrayList<JarFileWithPath> jarFiles;
+public class ClassLoaderResourceLocations<R extends HierarchicalResource> implements Serializable {
+    public final ArrayList<R> directories;
+    public final ArrayList<JarFileWithPath<R>> jarFiles;
     public final ArrayList<URL> unrecognizedUrls;
 
     public ClassLoaderResourceLocations(
-        ArrayList<FSResource> directories,
-        ArrayList<JarFileWithPath> jarFiles,
+        ArrayList<R> directories,
+        ArrayList<JarFileWithPath<R>> jarFiles,
         ArrayList<URL> unrecognizedUrls
     ) {
         this.directories = directories;
@@ -29,7 +31,7 @@ public class ClassLoaderResourceLocations implements Serializable {
     @Override public boolean equals(@Nullable Object o) {
         if(this == o) return true;
         if(o == null || getClass() != o.getClass()) return false;
-        final ClassLoaderResourceLocations that = (ClassLoaderResourceLocations)o;
+        final ClassLoaderResourceLocations<?> that = (ClassLoaderResourceLocations<?>)o;
         if(!directories.equals(that.directories)) return false;
         if(!jarFiles.equals(that.jarFiles)) return false;
         return unrecognizedUrls.equals(that.unrecognizedUrls);