+++ api/src/main/java/mb/resource/classloader/ClassLoaderToNativeResolver.java
@@ -1,10 +1,12 @@
 package mb.resource.classloader;
 
-import mb.resource.ReadableResource;
+import mb.resource.hierarchical.HierarchicalResource;
 import org.checkerframework.checker.nullness.qual.Nullable;
 
 import java.net.URL;
 
 public interface ClassLoaderToNativeResolver {
-    @Nullable ReadableResource toNativeResource(URL url);
+    @Nullable HierarchicalResource toNativeFile(URL url);
+
+    @Nullable HierarchicalResource toNativeDirectory(URL url);
 }