+++ api/src/main/java/mb/resource/classloader/JarFileWithPath.java
@@ -1,16 +1,18 @@
 package mb.resource.classloader;
 
+import mb.resource.ReadableResource;
 import mb.resource.fs.FSResource;
+import mb.resource.hierarchical.HierarchicalResource;
 import org.checkerframework.checker.nullness.qual.Nullable;
 
 import java.io.Serializable;
 import java.util.Objects;
 
-public class JarFileWithPath implements Serializable {
-    public final FSResource file;
+public class JarFileWithPath<R extends ReadableResource> implements Serializable {
+    public final R file;
     public final String path;
 
-    public JarFileWithPath(FSResource file, String path) {
+    public JarFileWithPath(R file, String path) {
         this.file = file;
         this.path = path;
     }
@@ -18,7 +20,7 @@ public class JarFileWithPath implements Serializable {
     @Override public boolean equals(@Nullable Object o) {
         if(this == o) return true;
         if(o == null || getClass() != o.getClass()) return false;
-        final JarFileWithPath that = (JarFileWithPath)o;
+        final JarFileWithPath<?> that = (JarFileWithPath<?>)o;
         return file.equals(that.file) && path.equals(that.path);
     }